+++ lib/src/classifier/_helpers/create_log_likelihood_optimizer.dart
@@ -51,6 +51,7 @@ LinearOptimizer createLogLikelihoodOptimizer(
     linkFunction: linkFunction,
     positiveLabel: positiveLabel,
     negativeLabel: negativeLabel,
+    dtype: dtype,
   );
   final normalizedLabels =
       normalizeClassLabels(labels, positiveLabel, negativeLabel);