+++ lib/src/cost_function/cost_function_factory.dart
@@ -1,6 +1,7 @@
 import 'package:ml_algo/src/cost_function/cost_function.dart';
 import 'package:ml_algo/src/cost_function/cost_function_type.dart';
 import 'package:ml_algo/src/link_function/link_function.dart';
+import 'package:ml_linalg/dtype.dart';
 
 abstract class CostFunctionFactory {
   CostFunction createByType(
@@ -8,5 +9,6 @@ abstract class CostFunctionFactory {
     LinkFunction? linkFunction,
     num? positiveLabel,
     num? negativeLabel,
+    DType dtype,
   });
 }