+++ lib/src/cost_function/cost_function_factory_impl.dart
@@ -4,6 +4,7 @@ import 'package:ml_algo/src/cost_function/cost_function_type.dart';
 import 'package:ml_algo/src/cost_function/least_square_cost_function.dart';
 import 'package:ml_algo/src/cost_function/log_likelihood_cost_function.dart';
 import 'package:ml_algo/src/link_function/link_function.dart';
+import 'package:ml_linalg/dtype.dart';
 
 class CostFunctionFactoryImpl implements CostFunctionFactory {
   const CostFunctionFactoryImpl();
@@ -14,6 +15,7 @@ class CostFunctionFactoryImpl implements CostFunctionFactory {
     LinkFunction? linkFunction,
     num? positiveLabel,
     num? negativeLabel,
+    DType dtype = DType.float32,
   }) {
     switch (type) {
       case CostFunctionType.logLikelihood:
@@ -34,6 +36,7 @@ class CostFunctionFactoryImpl implements CostFunctionFactory {
           linkFunction,
           positiveLabel,
           negativeLabel,
+          dtype,
         );
 
       case CostFunctionType.leastSquare: