+++ lib/src/cost_function/log_likelihood_cost_function.dart
@@ -2,17 +2,20 @@ import 'package:ml_algo/src/cost_function/cost_function.dart';
 import 'package:ml_algo/src/helpers/normalize_class_labels.dart';
 import 'package:ml_algo/src/helpers/validate_class_labels.dart';
 import 'package:ml_algo/src/link_function/link_function.dart';
-import 'package:ml_linalg/linalg.dart';
+import 'package:ml_linalg/dtype.dart';
+import 'package:ml_linalg/matrix.dart';
+import 'package:ml_linalg/vector.dart';
 
 class LogLikelihoodCostFunction implements CostFunction {
-  LogLikelihoodCostFunction(
-      this._linkFunction, this._positiveLabel, this._negativeLabel) {
+  LogLikelihoodCostFunction(this._linkFunction, this._positiveLabel,
+      this._negativeLabel, this._dtype) {
     validateClassLabels(_positiveLabel, _negativeLabel);
   }
 
   final LinkFunction _linkFunction;
   final num _positiveLabel;
   final num _negativeLabel;
+  final DType _dtype;
 
   @override
   double getCost(Matrix x, Matrix w, Matrix y) {
@@ -39,9 +42,9 @@ class LogLikelihoodCostFunction implements CostFunction {
   @override
   Matrix getHessian(Matrix x, Matrix w, Matrix y) {
     final prediction = _linkFunction.link(x * w).toVector();
-    final ones = Vector.filled(x.rowsNum, 1.0, dtype: x.dtype);
+    final ones = Vector.filled(x.rowsNum, 1.0, dtype: _dtype);
     final V = Matrix.diagonal((prediction * (ones - prediction)).toList(),
-        dtype: x.dtype);
+        dtype: _dtype);
 
     return x.transpose() * V * x;
   }