+++ lib/src/linear_optimizer/linear_optimizer_type_to_json_encoded_value.dart
@@ -1,4 +1,4 @@
-import 'package:ml_algo/ml_algo.dart';
+import 'package:ml_algo/src/linear_optimizer/linear_optimizer_type.dart';
 import 'package:ml_algo/src/linear_optimizer/linear_optimizer_type_json_encoded_values.dart';
 import 'package:quiver/collection.dart';