+++ lib/src/regressor/linear_regressor/linear_regressor_impl.dart
@@ -20,7 +20,9 @@ import 'package:ml_algo/src/regressor/linear_regressor/linear_regressor_constant
 import 'package:ml_algo/src/regressor/linear_regressor/linear_regressor_factory.dart';
 import 'package:ml_algo/src/regressor/linear_regressor/linear_regressor_json_keys.dart';
 import 'package:ml_dataframe/ml_dataframe.dart';
-import 'package:ml_linalg/linalg.dart';
+import 'package:ml_linalg/dtype.dart';
+import 'package:ml_linalg/matrix.dart';
+import 'package:ml_linalg/vector.dart';
 
 part 'linear_regressor_impl.g.dart';