+++ lib/src/tree_trainer/splitter/nominal_splitter/nominal_splitter.dart
@@ -1,5 +1,4 @@
 import 'package:ml_algo/src/tree_trainer/tree_node/tree_node.dart';
-import 'package:ml_linalg/linalg.dart';
 import 'package:ml_linalg/matrix.dart';
 
 abstract class NominalTreeSplitter {