+++ test/classifier/knn_classifier/knn_classifier_integration_test.dart
@@ -9,7 +9,11 @@ import 'package:ml_algo/src/knn_kernel/kernel_type_json_keys.dart';
 import 'package:ml_algo/src/knn_solver/knn_solver_constants.dart';
 import 'package:ml_algo/src/knn_solver/knn_solver_json_keys.dart';
 import 'package:ml_dataframe/ml_dataframe.dart';
-import 'package:ml_linalg/linalg.dart';
+import 'package:ml_linalg/distance.dart';
+import 'package:ml_linalg/distance_type_to_json.dart';
+import 'package:ml_linalg/dtype.dart';
+import 'package:ml_linalg/dtype_to_json.dart';
+import 'package:ml_linalg/matrix.dart';
 import 'package:test/test.dart';
 
 void main() {