+++ test/classifier/logistic_regressor/integration_test/logistic_regressor_fitting_integration_test.dart
@@ -6,7 +6,6 @@ import 'package:ml_algo/src/linear_optimizer/gradient_optimizer/learning_rate/le
 import 'package:ml_algo/src/linear_optimizer/linear_optimizer_type.dart';
 import 'package:ml_dataframe/ml_dataframe.dart';
 import 'package:ml_linalg/dtype.dart';
-import 'package:ml_linalg/linalg.dart';
 import 'package:test/test.dart';
 
 import '../../../helpers.dart';